1、peak值除以10以后的结果；1000次积分

+++ dealusb_msg.h
@@ -159,7 +159,7 @@ public slots:
 
 
     //校正
-    float pileUp_calibration(int,int);
+    float pileUp_calibration(int,float);
 
 
     //自动校正相关
